+++ k_core_stats/k_core_stat_main.cpp
@@ -7,7 +7,7 @@ July 2011*/
 #include <iomanip>
 #include <cstdlib>
 #include <ctime>
-#include "graph_lib_boost.hpp"
+#include "../lib/graph_lib_boost.hpp"
 #include <stdio.h>
 #include <sys/stat.h>
 #include <sys/types.h>
@@ -99,7 +99,7 @@ int main(int argc, char *argv[])
 
   int max_core = 0;
  
-  for(int i=0; i<core.size();i++)
+  for(size_t i = 0; i < core.size(); i++)
     {      
       if(core[i]>max_core)
 	max_core = core[i];
@@ -118,7 +118,7 @@ int main(int argc, char *argv[])
     {
       vector<float> core_dist(max_core+1,0);
   
-      for(int i=0;i<core.size();i++)
+      for(size_t i = 0; i < core.size(); i++)
 	core_dist[(core[i])]++;
 
       
@@ -133,7 +133,7 @@ int main(int argc, char *argv[])
       //     snapshot(Gfull,core,iter+1,filename);
       //   }
       float sum = 0;
-      for(int i=1;i<=max_core+1;i++)
+      for(int i = 1;i <= max_core + 1; i++)
 	{
 	  core_dist[i-1] = core_dist[i-1]/size;
 	  sum += core_dist[i-1];
@@ -144,7 +144,7 @@ int main(int argc, char *argv[])
 
       cout<<"Shell #\tEdges\tEdgesin\tAvg\tVar\t0\t1-5in\t1-5out\t5-10in\t5-10out\t10-20in\t10-20ou\t>20in\t>20ou\tdistr\n";
       cout<<fixed;
-      for(int i=0;i<core_stat.size();i++)
+      for(size_t i = 0; i < core_stat.size(); i++)
   	{
   	  vector<float> stat = core_stat[i];
   	  cout<<i<<": \t";
@@ -170,7 +170,7 @@ int main(int argc, char *argv[])
 
   string xtics = "";
 
-  for(int i=0;i<=max_core;++i)
+  for(int i = 0; i <= max_core; ++i)
     {
       if(i==max_core)
 	{
@@ -197,37 +197,5 @@ int main(int argc, char *argv[])
     xtic = "5";
     
   produce_plot(core_stat_file, plot_avg_output, columns_to_plot, label, "Average k-shell Edge Jump", "Start Shell", "Avg. shell Change",xstart,xtic,false);
-
-  // vector<int> cores_to_plot;
-  // cores_to_plot.push_back(1);
-  // cores_to_plot.push_back(5);
-  // cores_to_plot.push_back(10);
-  // cores_to_plot.push_back(20);
-  // cores_to_plot.push_back(40);
-  // cores_to_plot.push_back(80);
-
-  // vector<int> columns_to_plot;
-  // vector<string> label;
-  // for(int i=0;i<cores_to_plot.size();i++)
-  //   {
-  //     stringstream ss1;
-  //     ss1<<"Core "<<cores_to_plot[i];
-  //     columns_to_plot.push_back(cores_to_plot[i]+2);
-  //     label.push_back(ss1.str());
-  //   }
-
-  // string xtics = "\"0\" 0, \"1-5\" 1, \"5-10\" 2, \"10-20\" 3, \">20\" 4";
-
-  // string outfilename1 = prefix;
-  // outfilename1.append("_out_plot");
-
-  // string outfilename2 = prefix;
-  // outfilename2.append("_in_plot");
-
-  // produce_plot(filein, outfilename2,columns_to_plot,label,"Edges to k-core Interior", "Size of Jump", "Percent of Edges", xtics);
-
-  // produce_plot(fileout,outfilename1,columns_to_plot,label,"Edges to k-core Exterior", "Size of Jump", "Percent of Edges", xtics);
-
-
     
 }
